+++ server/matching/src/main/java/meet_at_mensa/matching/model/Match.java
@@ -1,4 +1,4 @@
-package meet_at_mensa.matching;
+package meet_at_mensa.matching.model;
 
 // import persistence tags
 import jakarta.persistence.Entity;
