+++ server/user/src/main/java/meet_at_mensa/user/controller/UserController.java
@@ -1,4 +1,4 @@
-package meet_at_mensa.user;
+package meet_at_mensa.user.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
 
@@ -9,6 +9,10 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.server.ResponseStatusException;
+
+import meet_at_mensa.user.model.User;
+import meet_at_mensa.user.repository.UserRepository;
+
 import org.springframework.http.HttpStatus;
 
 import java.util.UUID;
