+++ server/user/src/main/java/meet_at_mensa/user/model/Interest.java
@@ -1,4 +1,4 @@
-package meet_at_mensa.user;
+package meet_at_mensa.user.model;
 
 // import persistence tags
 import jakarta.persistence.Entity;
